Move to pkg/args CommandRunner
